Updated the knob to have correct smoothing groups. I likely exported it incorrectly at work, due to not having the export settings I had figured out there. Alternatively, the new version of Blender may have helped.
Remported the knob as a skeletal mesh (Unreal will just assign it a root bone at the origin, so I don't have to bother in Blender) and created a curve-based animation that rotates the knob. While the animation could have been done purely in blueprint, this approach should be more flexible since any animation is possible, and fairly easy to put in. For some reason Unreal wanted to update the materials as well.