Changed the interactor and explorer blueprints to allow for "deviation", that is, mouse movement while an interaction is taking place. In this implementation, for touching that means that you can hold down the left mouse button on a touchable interactable and move the mouse to send movement to that interactable. The interactable can then use this data to do something, for example animate. This is perfect for things like levers, sliders, or... knobs! The blueprint for the knob makes use of this functionality so you can click it and rotate it. The animation gives it the appearance that it snaps to (seven) notches in the knob, much like an old volume button.
The functionality also exists for hover interactions, which works much the same except mouse movement still moves the camera. This could be interesting for things that sense your hand or something.

This shows that you can build pretty cool stuff on top of the basic C++ plugin, combining blueprint subclassing with that interactor-to-interactable data flow. The blueprint can get a little complex depending on the functionality, but the simplicity of the plugin it keeps things very flexible.

But that's not all! The detector now checks the knobs to make sure you've entered the right code before it allows you to use the button! This means that this is the first real UnrealMYST puzzle!!! But what is the code? Maybe it's hidden somewhere where you can't normally see it... or just watch the (unlisted) YouTube walkthrough: https://www.youtube.com/watch?v=os2B2DOYwKU

All that remains now for the basic version of UnrealMYST to be finished is linking... and maybe some ages? :o