Implemented linking between ages through level transitions.
A link-in point blueprint has been added that can placed in an age and given a name and whether it is the default link-in point. Linking is then done by calling the LinkOut event on the explorer, naming an age and a link-in point in that age. If no age is specified, linking is done to the current age (something which is generally not possible in D'ni lore, but having a fallback is nice). If the named age does not exist, Unreal will attempt to load the default level (this is by Unreal design, and outside of what we can control). If the default level does not exist either, Unreal will terminate with an error. If the age exists but the link-in point does not, the default link-in point will be used. If there is no default link-in point, any link-in point will have to do. If there are no link-in points, the explorer will remain on the player start (an explorer will always spawn on the player start as per Unreal design, and is moved to the correct link-in point afterwards). If there are no player starts, Unreal will strongly warn you when building a level, and you're just being obtuse.

A link component is available for (for example) interactables to reduce the amount of blueprint necessary. It allows setting an age and a link-in point, and triggers the linking of the explorer (if it's the interactor) with these names when its Link event is called. The new linking book uses this component to link the explorer to the age and link-in point specified in the component (in the level instance) upon touch interaction. The test age has two link-in points, a default one and an alternative one that the linking book links to.

So far this is mostly how things worked in UDKMYST with some tweaks and changes here and there, but the link-in point business is implemented somewhat differently. Tthe link-in point to link to is stored in the game instance (called a journey, lol) and retrieved in the age that was linked to, to move the explorer to the correct position and rotation. This game instance can be customized or extended for more cross-age stuff, as needed by the writer. Don't forget to update the project configuration as necessary.

The new, custom link sound has also been added into Unreal, and the previous MYST link sound has been removed. It is Cyan property after all, and should not have been on this repository. Note that the link sound, nor any linking effect such as the traditional fade, have been implemented yet.

Finally, a note about input... in the original UDKMYST, cinematic mode was enabled during linking to prevent user input. Since we can now more easily control how the explorer handles input due to the use of blueprint, a simple boolean set to true during linking is used as a guard to prevent user interaction, similar to how the other booleans already work. However, note that only left mouse (touch interaction) is blocked during linking. Right and middle mouse still toggle the movement booleans, which are only read during the tick event, which in turned is guarded by the linking boolean. This allows the user to attempt to move while still linking, moving right after linking is completed. Otherwise, the user will have to let go of the mouse button and repress it, which can be annoying. Such cases are not as likely to occur with the left mouse button, unless the link-in point is directly in front of a touchable interactable that can be held. This latter case is not taken into account, to keep the blueprint significantly simpler and more readable.