Added screen fading when linking. The linking experience is now similar to the ones from the original MYST series. The fades are tuned to the linking sound, to give the best feel (that I could manage).
The fade is implemented as a HUD overlay, essentially hiding the scene behind a black rectangle. It uses a simple HUD material that has an opacity parameter.

The length of the linking process is controlled by 3 time variables, one for a pre-linking delay, one for the linking delay itself, and one for a post-linking delay. These are configurable in the explorer. The lengths of the sound and fade curves are not used to control the linking process, because while possible, it's not as flexible. It does mean that when editing a linking process, more values have to be changed (there is a link-in time of 3 seconds, a curve that is 3 seconds long, and a timeline that is 3 seconds long). But since Unreal already requires you to distribute some of these values, I figured one more couldn't hurt. Additionally could opt to mismatch some of these. For example, the link-in delay is only 3 seconds despite the sound being 7 seconds long, since most of those remaining 4 seconds is a fadey echo. But the link-out delay is 7 seconds, because otherwise the sound would cut off.

Linking in and out both have their own fade curves, and are as such controlled by different timelines. The HUD material is still the same one though, so technically this could cause a visua hiccup if both are activated in each other's ranges. Currently that's not possible from a user perspective due to the input locking, but technically it could be done by the level designer. Something to keep in mind.

When audio is added to an age, we may have to look into fading the master volume as well.