Added the ability to run by holding down the space bar (that is to say, it increases the current movement speed). There is a variable configured in the explorer whose value is multiplied with the base speed while the space bar is held down. Running is currently 2.5 times faster than exploring regularly, which feels accurate enough.
While this was in UDKMYST, I never added it to this version because I was aiming for a mouse-only control scheme (and perhaps to keep things lean and focused for the time being) and I ran out of mouse buttons. However, enabling it with the largest key on the keyboard is something your non-mouse hand can do that is not super complex, and as such this makes it ideal for the less gamery among us. But had I added it earlier, it would have saved me a lot of development time during testing XD

If we ignore all the overengineering and vehicle nonsense, I think this puts this version at feature parity with UDKMYST (except that this version has more features!). Time to write some ages!