Updated the descriptive book source with the required UV map and material stuff, and scaled it to its intended size. Imported and fleshed it out in Unreal similar to the linking book. It has the same setup, and also a parallax effect. Due to the angle of the panel, it looks different, but by changing the panel material so that the average color is no longer for the background, but the page color, it looks pretty ok. Also backported and of such upgrades into the linking book.
I also discovered that, YET AGAIN, the export/import process from Blender to Unreal was borked somewhere, as Unreal has +X as the forward axis (something I forgot). I was unable to get Blender to export the FBX is such a way that Unreal could import it and convert it to its own coordinate system, but I discovered that simply disabling Unreal attempts to do so let it import the FBX correctly. In the end I never had to change any of the Blender export settings. I suspect the auto-convert option was added fairly recently and is on by default, so that mucked up the process I had figured out. But now the linking book and the descriptive book are both +X forward, so they are now "correct" and don't need rotations anymore in the test level :P I also had to change the panel materials to account for the new transformation from Y/X in Unreal coordinates to X/Y in texture coordinates.