In this giant commit, I've reorganized the resources, both the source folder and the Unreal imports.
The "new" layout consists of:
- A folder for ages, with a folder for each age and their custom assets. There are currently two ages (Hub and D'nalsi), both of which are an empty level, but the Unreal project config has been updated to officially start on the Hub age now.
- A folder for generic purpose assets, currently containing the red button and knob from the test level.
- A folder for the engine, AKA UnrealMYST :P. Nothing changed in this folder on the Unreal side, but the resource source folders now reflect this layout.

There is also the Test folder on the Unreal side still. I'm not sure how much longer that is going to be around. It has no resource folder backing either. I moved the level into its own subfolder.

The resources folder can also contain additional folders for generic things like logos. It's not like it has to ONLY mirror Unreal, it's just that a flat list is going to be hard to work with in the long run.

I also "had to" update all the Unreal imports with the new source file paths. As a bonus, some paths were absolute but are now relative (I'm guessing due to Unreal version differences), so they should work across working copies.