Added a plane version of the grid block asset (a grid plane!) and updated the hub accordingly. It wastes less UV space on parts that will not be lit (more quality lightmaps!) and allows seeing through their backs in the level editor.
Combined with the grid material, and its ability to have instances with different textures, this basically mimics how BSP works/worked from a user's perspective. That's pretty cool, it makes it easier to reuse things in the editor, and you (as in, I) don't have to create the entire level in Blender like one normally would.

Consequently, I expect to simply use the grid assets to set up most of the hub, using with a custom set of textures (narble?).